😊 :: translate 데이터 분할 수정

+++ model/translate_T5.py
@@ -1,5 +1,6 @@
 # https://metamath1.github.io/blog/posts/gentle-t5-trans/gentle_t5_trans.html
 
+import os
 from datasets import load_dataset, Dataset, load_metric
 from transformers import AutoTokenizer, AutoModelForSeq2SeqLM, DataCollatorForSeq2Seq
 from transformers import Seq2SeqTrainingArguments, Seq2SeqTrainer
@@ -28,14 +29,35 @@ num_test = 10000
 en_ko_df_train = en_ko_df.iloc[:num_train]
 en_ko_df_valid = en_ko_df.iloc[num_train:num_train + num_valid]
 en_ko_df_test = en_ko_df.iloc[-num_test:]
-en_ko_df_train.to_csv("train.tsv", sep='\t', index=False)
-en_ko_df_valid.to_csv("valid.tsv", sep='\t', index=False)
-en_ko_df_test.to_csv("test.tsv", sep='\t', index=False)
+
+if os.path.exists("train.tsv"):
+  print("------------------------------")
+  print("train.tsv 존재")
+  print("------------------------------")
+else:
+  en_ko_df_train.to_csv("train.tsv", sep='\t', index=False)
+
+if os.path.exist("valid.tsv"):
+  print("------------------------------")
+  print("valid.tsv 존재")
+  print("------------------------------")
+else:
+  en_ko_df_valid.to_csv("valid.tsv", sep='\t', index=False)
+
+if os.path.exist("test.tsv"):
+  print("------------------------------")
+  print("test.tsv 존재")
+  print("------------------------------")
+else:
+  en_ko_df_test.to_csv("test.tsv", sep='\t', index=False)
 
 datafiles = {"train": "train.tsv", "valid": "valid.tsv", "test": "test.tsv"}
 dataset = load_dataset("csv", data_files=datafiles, delimiter="\t")
 
 device = 'cuda' if torch.cuda.is_available() else 'cpu'
+print("------------------------------")
+print(device)
+print("------------------------------")
 model_ckpt = "KETI-AIR/ke-t5-base"
 max_token_length = 64
